Move image function to edit
