+++ src/blocks/block-post-grid/components/inspector.js
@@ -2,15 +2,9 @@
  * Inspector Controls
  */
 
-import compact from 'lodash/compact';
-import map from 'lodash/map';
-import get from 'lodash/get';
-
 // Setup the block
 const { __ } = wp.i18n;
 const { Component } = wp.element;
-const { compose } = wp.compose;
-const { withSelect } = wp.data;
 
 // Import block components
 const {
@@ -32,25 +26,10 @@ const MAX_POSTS_COLUMNS = 4;
 /**
  * Create an Inspector Controls wrapper Component
  */
-class Inspector extends Component {
+export default class Inspector extends Component {
 
     constructor( props ) {
 		super( ...arguments );
-	}
-
-	getImageSizeOptions() {
-		const { imageSizes, image } = this.props;
-
-		return compact( map( imageSizes, ( { name, slug } ) => {
-			const sizeUrl = get( image, [ 'media_details', 'sizes', slug, 'source_url' ] );
-			if ( ! sizeUrl ) {
-				return null;
-			}
-			return {
-				value: sizeUrl,
-				label: name,
-			};
-		} ) );
 	}
 
 	render() {
@@ -105,13 +84,6 @@ class Inspector extends Component {
 
         // Check the post type
 		const isPost = attributes.postType === 'post';
-
-		// const getSettings = wp.data.select( 'core/editor' ).getEditorSettings();
-		// console.log( getSettings );
-
-		const imageSizeOptions = this.getImageSizeOptions();
-
-		console.log(imageSizeOptions);
 
 		return (
             <InspectorControls>
@@ -271,18 +243,3 @@ class Inspector extends Component {
 		);
 	}
 }
-
-export default compose( [
-	withSelect( ( select, props ) => {
-		const { getMedia } = select( 'core' );
-		const { getEditorSettings } = select( 'core/editor' );
-		const { id } = props.attributes;
-		const { imageSizes } = getEditorSettings();
-
-		return {
-			image: 8187 ? getMedia( 8187 ) : null,
-			imageSizes,
-		};
-	} ),
-] )( Inspector );
-
